Smart Energy Monitor/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
numpy==1.24.3
scikit-learn==1.3.2
pandas==2.0.3
google-generativeai==0.3.2

